Graph So Far So Good
So far so good the adjacentcy matrix used to represent the graph seems
to be working. It needs to be edited to include database information
once the db is up and running and the topological sort method needs to
be implemented alongside the time filter. This project is almost done